% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numbers.R
\name{add_numbers}
\alias{add_numbers}
\title{Add together two numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numbers as input and returns their sum.
}
\examples{
add_numbers(1, 2)
}
