% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{distinct_voting_places}
\alias{distinct_voting_places}
\title{Distinct voting places}
\format{
A data frame with 3,425 rows and 8 columns
}
\source{
\url{http://www.electionresults.govt.nz/electionresults_2014/e9/html/statistics.html} (and similar for earlier years) 
for the voting place locations.
\url{http://www.stats.govt.nz/browse_for_stats/Maps_and_geography/Geographic-areas/digital-boundary-files.aspx} 
for the 2014 shapefiles for Regional Council, Territorial Authority, and Area Unit.  See
\url{https://github.com/ellisp/nzelect/tree/master/prep} for the code linking the two.
}
\usage{
distinct_voting_places
}
\description{
De-duplicated voting places from 2008 to 2017 with unique IDs allocated to them
}
\details{
The locations listed in the Electoral Commissions' various lists of voting places have been matched 
for likely duplicates and unique ids assigned to them.
}
\seealso{
\code{\link{nzge}}
}
\keyword{datasets}
