% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2.R
\name{fun2}
\alias{fun2}
\title{fun2}
\usage{
fun2(x = 1)
}
\arguments{
\item{x}{number larger than 0}
}
\value{
fun2 computes common (i.e., base 10) logarithms
}
\description{
fun2 computes common (i.e., base 10) logarithms.
}
\details{
input x should be a positive number
}
