% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print a Hello World message}
\usage{
hello()
}
\value{
None. This function is called for its side effect.
}
\description{
This function prints "Hello, world!" to the console.
}
\examples{
hello()
}
