% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_test.R
\name{median_test}
\alias{median_test}
\title{Monte-Carlo test of differences between medians}
\usage{
median_test(x, fac, n = 999, ...)
}
\arguments{
\item{x}{A vector of numeric values}

\item{fac}{A factor defining 2 groups}

\item{n}{The number of observations}
}
\description{
This is a test we developed with DIDE's MRes.
}
\examples{

## CREATE DATA ##
set.seed(1)
titer <- c(rexp(100, rate = 0.002), rexp(20, rate = 0.0005), # control
        rexp(55, rate = 0.006), rexp(15, rate = 0.0005)) # treatment

type <- factor(rep(c("control","treatment"), c(120,70)))

## make a data frame
VL <- data.frame(titer = titer, type = type)
}
