% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzelect.R
\docType{package}
\name{nzelect-package}
\alias{nzelect}
\alias{nzelect-package}
\title{New Zealand Election Data}
\description{
Convenient access to New Zealand election results by voting place.  
Voting places have been matched to Regional Council,
Territorial Authority, and Area Unit, to facilitate matching with additional data.
Opinion polls since 2002 and some convenience analytical function are also supplied.
}
\author{
\strong{Maintainer}: Peter Ellis \email{peter.ellis2013nz@gmail.com}

}
