% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parties_doc.R
\docType{data}
\name{parties_df}
\alias{parties_df}
\title{New Zealand political parties}
\format{
A data frame of colours, shades, short and long names
}
\source{
\url{https://en.wikipedia.org/wiki/Wikipedia:Index_of_New_Zealand_political_party_meta_attributes}
}
\usage{
parties_df
}
\description{
Metadata associated with New Zealand political parties in the early 21st century
}
\keyword{datasets}
