% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parties_doc.R
\docType{data}
\name{parties_v}
\alias{parties_v}
\title{New Zealand political party colours}
\format{A named vector of colours}
\source{
\url{https://en.wikipedia.org/wiki/Wikipedia:Index_of_New_Zealand_political_party_meta_attributes}
}
\usage{
parties_v
}
\description{
Colours associated with New Zealand political parties in the early 21st century
}
\keyword{datasets}
