% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg1-package.R
\docType{package}
\name{pkg1-package}
\alias{pkg1}
\alias{pkg1-package}
\title{pkg1: Example Package One}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Author Name \email{example@example.com}

}
\keyword{internal}
