% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{plusmat}
\alias{plusmat}
\title{Sum of two matrices}
\usage{
plusmat(A, B)
}
\arguments{
\item{A}{\assert{is.matrix(.) && identical(dim(A), dim(B))}
The first matrix.}

\item{B}{\assert{is.matrix(.) && identical(dim(A), dim(B))}
The second matrix.}
}
\value{
Their sum.
}
\description{
Sum of two matrices
}

