% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{prefix}
\alias{prefix}
\alias{suffix}
\title{Prefix of a string}
\usage{
prefix(str, len)

suffix(str, len)
}
\arguments{
\item{str}{\assert{is.character} Character vector.}

\item{len}{\assert{is.integer} Integer vector.}
}
\value{
Prefix is string, of prescribed length.
}
\description{
Prefix of a string

Suffix of a string
}

