% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{prefix2}
\alias{prefix2}
\title{Return the prefixes of specified number of characters}
\usage{
prefix2(str, len)
}
\arguments{
\item{str}{\coerce{as.character(str)} Character vector.}

\item{len}{\coerce{as.integer(len)} Integer vector.}
}
\value{
Prefix is string, of prescribed length.
}
\description{
Return the prefixes of specified number of characters
}

