% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pkg1-package.R
\docType{package}
\name{Pkg1-package}
\alias{Pkg1}
\alias{Pkg1-package}
\title{Pkg1: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangKai7kkw/Pkg1}
  \item \url{https://wangkai7kkw.github.io/Pkg1/}
  \item Report bugs at \url{https://github.com/WangKai7kkw/Pkg1/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
